Form request for new reservation, fixed UI
The new reservation request now has its own request class with the required validation. It validates all the data. It also checks if the selected gazebo was booked on the same day, if the day was disabled or if the gazebo was disabled for that date, while the user was filling the reservation form.

Added a note to indicate which kind of reservations ( Dinner or Bed ) are visible in Today's, Weekly, and Monthly View ( all both Large Screens and mobiles ).

Made some minor UI fixes ( paddings ).

When a reservation is edited in Today's view, it should now fetch the correct type of reservations to re-populate the list. ( was always fetching Dinner reservations., no matter what ).